float to double

+++ datasets/llff.py
@@ -190,7 +190,7 @@ class LLFFDataset(Dataset):
 
         # Step 1: rescale focal length according to training resolution
         # H, W, self.focal = poses[0, :, -1] # original intrinsics, same for all images
-        H, W, self.focal = hwf_cxcy[0], hwf_cxcy[1], hwf_cxcy[2] # original intrinsics, same for all images
+        H, W, self.focal = hwf_cxcy[0].double(), hwf_cxcy[1].double(), hwf_cxcy[2].double() # original intrinsics, same for all images
         assert H*self.img_wh[0] == W*self.img_wh[1], \
             f'You must set @img_wh to have the same aspect ratio as ({W}, {H}) !'
